pyproject.toml
[project]
name = "un-system-chart-navigator"
version = "0.1.0"
description = "Interactively navigate the United Nations System"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "black>=25.1.0",
    "ipykernel>=6.29.5",
    "isort>=6.0.1",
    "pandas>=3.0.0",
    "playwright>=1.55.0",
    "psycopg2-binary>=2.9.11",
    "pyairtable>=3.2.0",
    "pyarrow>=23.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "tqdm>=4.67.1",
]
